feat(2025-3d-terrain-demo): refactor cesium viewer and state management
